BUGFIX: Soft Reset Button; BUGFIX: setBrightness() bounds checking in MicroBitDisplay; BUGFIX: Moved to Galois LFSR based Random Number Generator

+++ inc/MicroBitDisplay.h
@@ -383,7 +383,7 @@ public:
       * uBit.display.setBrightness(255); //max brightness
       * @endcode
       */
-    void setBrightness(uint8_t b);
+    void setBrightness(int b);
 
     /**
       * Fetches the current brightness of this display.
